Continued work on GraphKey
* Implemented some base Entity nodes
* Implemented more versions of math nodes
* Implemented feedback for incorrect link creations
* Implemented a simple scene manager

+++ Volt/Sandbox/src/Sandbox/Window/MaterialEditorPanel.cpp
@@ -358,7 +358,7 @@ void MaterialEditorPanel::UpdateSubMaterials()
 			const auto& subMaterials = mySelectedMaterial->GetSubMaterials();
 			for (auto& [index, material] : subMaterials)
 			{
-				ImGui::PushID((uint32_t)material->GetName().c_str());
+				ImGui::PushID(material->GetName().c_str());
 
 				constexpr float thumbnailSize = 85.f;
 				const ImVec2 itemSize = AssetBrowserUtilities::GetBrowserItemSize(thumbnailSize);
